moves packages to repository
