+++ artemis/artemis-core/src/main/java/org/eclipse/jnosql/artemis/reflection/repository/DefaultDynamicExecutorQueryConverter.java
@@ -12,10 +12,12 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.artemis.reflection;
+package org.eclipse.jnosql.artemis.reflection.repository;
 
 import jakarta.nosql.mapping.DynamicQueryException;
 import jakarta.nosql.mapping.Page;
+import org.eclipse.jnosql.artemis.reflection.repository.DynamicExecutorQueryConverter;
+import org.eclipse.jnosql.artemis.reflection.repository.DynamicReturn;
 
 import java.util.LinkedList;
 import java.util.List;
