+++ artemis/artemis-core/src/main/java/org/eclipse/jnosql/artemis/reflection/repository/DynamicReturnConverter.java
@@ -12,9 +12,10 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.artemis.reflection;
+package org.eclipse.jnosql.artemis.reflection.repository;
 
 import jakarta.nosql.mapping.PreparedStatement;
+import org.eclipse.jnosql.artemis.reflection.RepositoryReflectionUtils;
 
 import java.lang.reflect.Method;
 import java.util.Map;
