+++ artemis/artemis-core/src/test/java/org/eclipse/jnosql/artemis/reflection/repository/DynamicQueryMethodReturnTest.java
@@ -12,13 +12,14 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.artemis.reflection;
+package org.eclipse.jnosql.artemis.reflection.repository;
 
 import jakarta.nosql.NonUniqueResultException;
 import jakarta.nosql.mapping.Param;
 import jakarta.nosql.mapping.PreparedStatement;
 import jakarta.nosql.mapping.Query;
 import jakarta.nosql.mapping.Repository;
+import org.eclipse.jnosql.artemis.reflection.repository.DynamicQueryMethodReturn;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
