+++ artemis/artemis-core/src/test/java/org/eclipse/jnosql/artemis/reflection/repository/DynamicReturnTypeTest.java
@@ -12,10 +12,11 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.artemis.reflection;
+package org.eclipse.jnosql.artemis.reflection.repository;
 
 import jakarta.nosql.mapping.DynamicQueryException;
 import jakarta.nosql.mapping.Page;
+import org.eclipse.jnosql.artemis.reflection.repository.DynamicReturnType;
 import org.junit.jupiter.api.Test;
 
 import java.util.Collection;
